container-validator/src/requirements.txt
psycopg2
environs
pydantic
python-dotenv
pandas
envs
schedule
